refactor: merge all global settings into root config
