+++ tests/test_download_client.py
@@ -15,13 +15,13 @@ from cyberdrop_dl.manager import Manager
 )
 def test_chunk_size_is_never_greater_that_speed_limit(manager: Manager, limit: int) -> None:
     max_expected = 1024 * 1024 * 10
-    limit = manager.config.global_settings.rate_limiting_options.download_speed_limit
+    limit = manager.config.rate_limiting_options.download_speed_limit
     assert limit == 0
     client = DownloadClient(manager)
     assert client.chunk_size != limit
     assert client.chunk_size == max_expected
 
-    manager.config.global_settings.rate_limiting_options.download_speed_limit = limit
+    manager.config.rate_limiting_options.download_speed_limit = limit
     client = DownloadClient(manager)
     assert client.chunk_size == min(limit or max_expected, max_expected)
